Refactors SpanAttributeSchema into tracing to be used both inside and outside of contrib for internal tags
