+++ lib/datadog/tracing/contrib/aws/configuration/settings.rb
@@ -1,6 +1,6 @@
 require_relative '../../configuration/settings'
 require_relative '../ext'
-require_relative '../../span_attribute_schema'
+require_relative '../../../span_attribute_schema'
 
 module Datadog
   module Tracing
@@ -27,7 +27,7 @@ module Datadog
 
             option :service_name do |o|
               o.default do
-                Contrib::SpanAttributeSchema.fetch_service_name(
+                SpanAttributeSchema.fetch_service_name(
                   Ext::ENV_SERVICE_NAME,
                   Ext::DEFAULT_PEER_SERVICE_NAME
                 )
