+++ lib/datadog/tracing/contrib/aws/instrumentation.rb
@@ -1,7 +1,7 @@
 require_relative '../../metadata/ext'
 require_relative '../analytics'
 require_relative 'ext'
-require_relative '../span_attribute_schema'
+require_relative '../../span_attribute_schema'
 
 module Datadog
   module Tracing
@@ -38,7 +38,7 @@ module Datadog
             span.set_tag(Tracing::Metadata::Ext::TAG_OPERATION, Ext::TAG_OPERATION_COMMAND)
 
             # Tag as an external peer service
-            if Contrib::SpanAttributeSchema.default_span_attribute_schema?
+            if SpanAttributeSchema.default_span_attribute_schema?
               span.set_tag(Tracing::Metadata::Ext::TAG_PEER_SERVICE, span.service)
               span.set_tag(Tracing::Metadata::Ext::TAG_PEER_HOSTNAME, context.safely(:host))
             end
